requirements.txt
numpy~=2.1.3
Flask~=3.1.1
tensorflow~=2.19.0
pillow~=11.2.1